[MOD] If passed arguments path is not invalid, don't remove it

+++ remove.go
@@ -9,6 +9,10 @@ import (
 )
 
 func cleanDir(path string) {
+	if !contains(path) {
+		fmt.Printf("Invalid path: %s", path)
+		return
+	}
 	os.RemoveAll(path)
 	unregister(path)
 }
